Update the UnificationBinaryOp so it inherits from binaryExpr
