+++ src/main/java-gen/dev/monogon/cue/lang/psi/CueField.java
@@ -16,7 +16,7 @@ public interface CueField extends CueDeclaration {
 
   @NotNull
   List<CueAttribute> getAttributeList();
-  @Nullable
-  ItemPresentation getPresentation();
+
+  @Nullable ItemPresentation getPresentation();
 
 }
