+++ src/main/java-gen/dev/monogon/cue/lang/psi/CueVisitor.java
@@ -157,7 +157,7 @@ public class CueVisitor extends PsiElementVisitor {
   }
 
   public void visitUnificationBinaryExpr(@NotNull CueUnificationBinaryExpr o) {
-    visitExpression(o);
+    visitBinaryExpr(o);
   }
 
   public void visitAttrTokens(@NotNull CueAttrTokens o) {
