+++ src/main/java-gen/dev/monogon/cue/lang/psi/impl/CueUnificationBinaryExprImpl.java
@@ -10,7 +10,7 @@ import com.intellij.psi.util.PsiTreeUtil;
 import static dev.monogon.cue.lang.CueTypes.*;
 import dev.monogon.cue.lang.psi.*;
 
-public class CueUnificationBinaryExprImpl extends CueExpressionImpl implements CueUnificationBinaryExpr {
+public class CueUnificationBinaryExprImpl extends CueBinaryExprImpl implements CueUnificationBinaryExpr {
 
   public CueUnificationBinaryExprImpl(@NotNull ASTNode node) {
     super(node);
@@ -25,26 +25,6 @@ public class CueUnificationBinaryExprImpl extends CueExpressionImpl implements C
   public void accept(@NotNull PsiElementVisitor visitor) {
     if (visitor instanceof CueVisitor) accept((CueVisitor)visitor);
     else super.accept(visitor);
-  }
-
-  @Override
-  @NotNull
-  public List<CueExpression> getExpressionList() {
-    return PsiTreeUtil.getChildrenOfTypeAsList(this, CueExpression.class);
-  }
-
-  @Override
-  @NotNull
-  public CueExpression getLeft() {
-    List<CueExpression> p1 = getExpressionList();
-    return p1.get(0);
-  }
-
-  @Override
-  @Nullable
-  public CueExpression getRight() {
-    List<CueExpression> p1 = getExpressionList();
-    return p1.size() < 2 ? null : p1.get(1);
   }
 
 }
